Bug 26884 - Part 1: Move show_torbrowser_manual and get_general_useragent_locale to utils
On Desktop, show_torbrowser_manual is a torbutton.js method and it
is called by other components using the window.opener.

However, on mobile, the security settings is opened in a tab instead
of a modal. Thus the window.opener is nil.

This patch moves the show_torbrowser_manual to util.js allowing it
to be accessible from anywhere.